Restrict mathching for package versions in package.json dependencies
The search string used to match dependencies and devDependencies is too
broad and captures other lines that contain debug switches like
    ''-- "debug", "log"'
because one of the affected packages is named debug. Configuration lines
like
     "supports-color": { "optional": true
are matched as well. This patch extends the requirements for a match by
including the colon separator.